Moved each cli command into own folder/package for easier management
